Fix forge update script not creating server.jar symlink if nonexistent

+++ update-forge.sh
@@ -14,11 +14,13 @@ wget $url
 java -jar forge-${version}-installer.jar --installServer
 rm -vf forge-${version}-installer.jar{,.log}
 
-serverJar=$(readlink server.jar)
-newJar="forge-${version}-universal.jar"
-
-if [ -n "$serverJar" ] && [[ "$serverJar" != minecraft* ]] && [ "$serverJar" != "$newJar" ]; then
-    rm -vf $serverJar
+if [ -e server.jar ]; then
+    serverJar=$(readlink server.jar)
+    newJar="forge-${version}-universal.jar"
+    
+    if [ -n "$serverJar" ] && [[ "$serverJar" != minecraft* ]] && [ "$serverJar" != "$newJar" ]; then
+        rm -vf $serverJar
+    fi
 fi
 
 ln -vfs forge-${version}-universal.jar server.jar
